requirements.txt
arcade==2.6.17
Pillow==10.0.0
pymunk==6.5.1
asyncio==3.4.3
websockets==11.0.3
jsonschema==4.17.3
numpy==1.24.3